Revert "[TT-2539] added access/transaction logs (#6354)"
This reverts commit 6f6190010d5c8b4e48ae313991201d1584e3861e.
