+++ cli/linter/schema.json
@@ -648,15 +648,6 @@
       "type": "string",
       "enum": ["", "standard", "json"]
     },
-    "access_logs": {
-      "type": ["object", "null"],
-      "additionalProperties": false,
-      "properties": {
-        "enabled": {
-          "type": "boolean"
-        }
-      }
-    },
     "enable_http_profiler": {
       "type": "boolean"
     },
