+++ gateway/handler_error.go
@@ -311,13 +311,6 @@ func (e *ErrorHandler) HandleError(w http.ResponseWriter, r *http.Request, errMs
 			log.WithError(err).Error("could not store analytic record")
 		}
 	}
-
-	// Print the transaction logs for error situations if enabled. Success transaction
-	// logs will be handled by the "handler_success.go"
-	if e.Spec.GlobalConfig.AccessLogs.Enabled {
-		e.recordAccessLog(r, response, nil)
-	}
-
 	// Report in health check
 	reportHealthValue(e.Spec, BlockedRequestLog, "-1")
 }
