+++ internal/httputil/Taskfile.yml
@@ -3,14 +3,13 @@ version: "3"
 
 vars:
   testArgs: -v
-  coverpkg: ./...,github.com/TykTechnologies/tyk/internal/httputil/...
 
 tasks:
   test:
     desc: "Run tests (requires redis)"
     cmds:
       - task: fmt
-      - go test {{.testArgs}} -count=1 -cover -coverprofile=rate.cov {{.coverpkg}} ./...
+      - go test {{.testArgs}} -count=1 -cover -coverprofile=rate.cov -coverpkg=./... ./...
 
   bench:
     desc: "Run benchmarks"
